config/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
cloudinary==1.31.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.2.0
Django==4.1.5
django-cors-headers==3.13.0
django-import-export==3.0.2
django-ip-geolocation==1.6.1
django-on-heroku==1.1.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.3.3
geoip2==4.6.0
gunicorn==20.1.0
idna==3.4
MarkupPy==1.14
maxminddb==2.2.0
multidict==6.0.4
numpy==1.24.1
odfpy==1.4.1
opencv-contrib-python==4.7.0.68
openpyxl==3.0.10
Pillow==9.4.0
psycopg2-binary==2.9.5
pypng==0.20220715.0
python-decouple==3.7
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
qrcode==7.4
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
typing_extensions==4.4.0
urllib3==1.26.14
whitenoise==6.3.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
